Simplification of DMRGateway.ini
This configuration assigns TS1 to DMRPlus/IPSC and TS2 to Brandmeister. DMRPlus Reflectors are on TS1 TG9. TS1 has to be empty in Brandmeister selfcare. In my opinion this is the simplest way use DMRGateway. Many configurations I encountered are using complex Talkgroup "NAT", so before I was able to make a contact I always had to search for a documentation of the one specific repeater I am using. In this configuration a single call to 5000 or the RPT-Command at 262994 tells me everything I need to know.