set INADDR_ANY/IN6ADDR_ANY_INIT address string to m_socket
when CUDPSocket::open() is called with m_address (in CUDPSocket) is nothing,
IPv4 socket is created by "0.0.0.0" (INADDR_ANY) address.

This causes a bug that DMRGateway cannot connect to reflector on IPv6.

Avoid this problem, set IP address string to m_socket (in CDMRNetwork) before
calling CUDPSocket::open().

This is a workaround, I think there is better idea.